refactor: 기존 프로젝트 구조를 패키지에 따라 분리
