+++ src/main/java/nextstep/auth/AuthenticationPrincipalArgumentResolver.java
@@ -10,14 +10,12 @@ import java.util.Optional;
 
 public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
 
-    // resolveArgument 메서드가 동작하는 조건을 정의하는 메서드
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
         // 파라미터 중 @AuthenticationPrincipal이 붙은 경우 동작하게 설정
         return parameter.hasParameterAnnotation(AuthenticationPrincipal.class);
     }
 
-    // supportsParameter가 true인 경우 동작하는 메서드
     @Override
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         String authorization = webRequest.getHeader("authorization");
