+++ src/main/java/nextstep/auth/controller/AuthController.java
@@ -1,0 +1,27 @@
+package nextstep.auth.controller;
+
+import nextstep.auth.dto.TokenRequest;
+import nextstep.auth.dto.TokenResponse;
+import nextstep.auth.service.AuthService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class AuthController {
+
+    private final AuthService authService;
+
+    @Autowired
+    public AuthController(AuthService authService) {
+        this.authService = authService;
+    }
+
+    @PostMapping("/login/token")
+    public ResponseEntity<TokenResponse> login(@RequestBody TokenRequest request) {
+
+        return ResponseEntity.ok(authService.login(request));
+    }
+}
