+++ src/main/java/nextstep/auth/dto/TokenRequest.java
@@ -1,0 +1,23 @@
+package nextstep.auth.dto;
+
+public class TokenRequest {
+
+    private String username;
+    private String password;
+
+    public TokenRequest() {
+    }
+
+    public TokenRequest(String username, String password) {
+        this.username = username;
+        this.password = password;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
