+++ src/main/java/nextstep/auth/dto/TokenResponse.java
@@ -1,0 +1,17 @@
+package nextstep.auth.dto;
+
+public class TokenResponse {
+
+    public String accessToken;
+
+    public TokenResponse() {
+    }
+
+    public TokenResponse(String accessToken) {
+        this.accessToken = accessToken;
+    }
+
+    public String getAccessToken() {
+        return accessToken;
+    }
+}
