--- src/main/java/nextstep/member/Member.java
@@ -1,48 +1,0 @@
-package nextstep.member;
-
-public class Member {
-    private Long id;
-    private String username;
-    private String password;
-    private String name;
-    private String phone;
-
-    public Member(Long id, String username, String password, String name, String phone) {
-        this.id = id;
-        this.username = username;
-        this.password = password;
-        this.name = name;
-        this.phone = phone;
-    }
-
-    public Member(String username, String password, String name, String phone) {
-        this.username = username;
-        this.password = password;
-        this.name = name;
-        this.phone = phone;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public boolean checkWrongPassword(String password) {
-        return !this.password.equals(password);
-    }
-}
