--- src/main/java/nextstep/member/MemberController.java
@@ -1,29 +1,0 @@
-package nextstep.member;
-
-import nextstep.auth.AuthenticationPrincipal;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
-
-@RestController
-@RequestMapping("/members")
-public class MemberController {
-    private MemberService memberService;
-
-    public MemberController(MemberService memberService) {
-        this.memberService = memberService;
-    }
-
-    @PostMapping
-    public ResponseEntity createMember(@RequestBody MemberRequest memberRequest) {
-        Long id = memberService.create(memberRequest);
-        return ResponseEntity.created(URI.create("/members/" + id)).build();
-    }
-
-    @GetMapping("/me")
-    public ResponseEntity<Member> me(@AuthenticationPrincipal String principal) {
-        Member member = memberService.findById(Long.parseLong(principal));
-        return ResponseEntity.ok(member);
-    }
-}
