--- src/main/java/nextstep/member/MemberService.java
@@ -1,20 +1,0 @@
-package nextstep.member;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class MemberService {
-    private MemberDao memberDao;
-
-    public MemberService(MemberDao memberDao) {
-        this.memberDao = memberDao;
-    }
-
-    public Long create(MemberRequest memberRequest) {
-        return memberDao.save(memberRequest.toEntity());
-    }
-
-    public Member findById(Long id) {
-        return memberDao.findById(id);
-    }
-}
