+++ src/main/java/nextstep/member/dao/MemberDao.java
@@ -1,0 +1,63 @@
+package nextstep.member.dao;
+
+import nextstep.member.entity.Member;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Component;
+
+import java.sql.PreparedStatement;
+import java.util.Optional;
+
+@Component
+public class MemberDao {
+
+    public final JdbcTemplate jdbcTemplate;
+
+    private final RowMapper<Member> rowMapper = (resultSet, rowNum) -> new Member(
+            resultSet.getLong("id"),
+            resultSet.getString("username"),
+            resultSet.getString("password"),
+            resultSet.getString("name"),
+            resultSet.getString("phone")
+    );
+
+    @Autowired
+    public MemberDao(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public Long save(Member member) {
+        String sql = "INSERT INTO member (username, password, name, phone) VALUES (?, ?, ?, ?);";
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+
+        jdbcTemplate.update(connection -> {
+            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
+            ps.setString(1, member.getUsername());
+            ps.setString(2, member.getPassword());
+            ps.setString(3, member.getName());
+            ps.setString(4, member.getPhone());
+            return ps;
+
+        }, keyHolder);
+
+        return keyHolder.getKey().longValue();
+    }
+
+    public Member findById(Long id) {
+        String sql = "SELECT id, username, password, name, phone from member where id = ?;";
+        return jdbcTemplate.queryForObject(sql, rowMapper, id);
+    }
+
+    public Member findByUsername(String username) {
+        String sql = "SELECT id, username, password, name, phone from member where username = ?;";
+        return jdbcTemplate.queryForObject(sql, rowMapper, username);
+    }
+
+    public Optional<Member> findByUsernameAndPassword(String username, String password) {
+        String sql = "SELECT id, username, password, name, phone from member where username = ? and password = ?;";
+        return Optional.ofNullable(jdbcTemplate.queryForObject(sql, rowMapper, username, password));
+    }
+}
