+++ src/main/java/nextstep/reservation/dto/ReservationRequest.java
@@ -1,0 +1,17 @@
+package nextstep.reservation.dto;
+
+public class ReservationRequest {
+
+    private Long scheduleId;
+
+    public ReservationRequest() {
+    }
+
+    public ReservationRequest(Long scheduleId) {
+        this.scheduleId = scheduleId;
+    }
+
+    public Long getScheduleId() {
+        return scheduleId;
+    }
+}
