+++ src/main/java/nextstep/schedule/dto/ScheduleRequest.java
@@ -1,0 +1,43 @@
+package nextstep.schedule.dto;
+
+import nextstep.schedule.entity.Schedule;
+import nextstep.theme.entity.Theme;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+public class ScheduleRequest {
+
+    private Long themeId;
+    private String date;
+    private String time;
+
+    public ScheduleRequest() {
+    }
+
+    public ScheduleRequest(Long themeId, String date, String time) {
+        this.themeId = themeId;
+        this.date = date;
+        this.time = time;
+    }
+
+    public Long getThemeId() {
+        return themeId;
+    }
+
+    public String getDate() {
+        return date;
+    }
+
+    public String getTime() {
+        return time;
+    }
+
+    public Schedule toEntity(Theme theme) {
+        return new Schedule(
+                theme,
+                LocalDate.parse(this.date),
+                LocalTime.parse(this.time)
+        );
+    }
+}
