+++ src/main/java/nextstep/schedule/service/ScheduleService.java
@@ -1,0 +1,37 @@
+package nextstep.schedule.service;
+
+import nextstep.schedule.dao.ScheduleDao;
+import nextstep.schedule.dto.ScheduleRequest;
+import nextstep.schedule.entity.Schedule;
+import nextstep.theme.dao.ThemeDao;
+import nextstep.theme.entity.Theme;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class ScheduleService {
+
+    private final ScheduleDao scheduleDao;
+    private final ThemeDao themeDao;
+
+    @Autowired
+    public ScheduleService(ScheduleDao scheduleDao, ThemeDao themeDao) {
+        this.scheduleDao = scheduleDao;
+        this.themeDao = themeDao;
+    }
+
+    public Long create(ScheduleRequest scheduleRequest) {
+        Theme theme = themeDao.findById(scheduleRequest.getThemeId());
+        return scheduleDao.save(scheduleRequest.toEntity(theme));
+    }
+
+    public List<Schedule> findByThemeIdAndDate(Long themeId, String date) {
+        return scheduleDao.findByThemeIdAndDate(themeId, date);
+    }
+
+    public void deleteById(Long id) {
+        scheduleDao.deleteById(id);
+    }
+}
