--- src/main/java/nextstep/theme/ThemeService.java
@@ -1,32 +1,0 @@
-package nextstep.theme;
-
-import nextstep.support.NotExistEntityException;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class ThemeService {
-    private ThemeDao themeDao;
-
-    public ThemeService(ThemeDao themeDao) {
-        this.themeDao = themeDao;
-    }
-
-    public Long create(ThemeRequest themeRequest) {
-        return themeDao.save(themeRequest.toEntity());
-    }
-
-    public List<Theme> findAll() {
-        return themeDao.findAll();
-    }
-
-    public void delete(Long id) {
-        Theme theme = themeDao.findById(id);
-        if (theme == null) {
-            throw new NotExistEntityException();
-        }
-
-        themeDao.delete(id);
-    }
-}
