+++ src/main/java/nextstep/theme/controller/ThemeController.java
@@ -1,0 +1,42 @@
+package nextstep.theme.controller;
+
+import nextstep.theme.dto.ThemeRequest;
+import nextstep.theme.entity.Theme;
+import nextstep.theme.service.ThemeService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+@RequestMapping("/themes")
+public class ThemeController {
+
+    private final ThemeService themeService;
+
+    @Autowired
+    public ThemeController(ThemeService themeService) {
+        this.themeService = themeService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Void> createTheme(@RequestBody ThemeRequest themeRequest) {
+        Long id = themeService.create(themeRequest);
+        return ResponseEntity.created(URI.create("/themes/" + id)).build();
+    }
+
+    @GetMapping
+    public ResponseEntity<List<Theme>> showThemes() {
+        List<Theme> results = themeService.findAll();
+        return ResponseEntity.ok().body(results);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteTheme(@PathVariable Long id) {
+        themeService.delete(id);
+
+        return ResponseEntity.noContent().build();
+    }
+}
