+++ src/main/java/nextstep/theme/entity/Theme.java
@@ -1,0 +1,41 @@
+package nextstep.theme.entity;
+
+public class Theme {
+
+    private Long id;
+    private String name;
+    private String desc;
+    private int price;
+
+    public Theme() {
+    }
+
+    public Theme(Long id, String name, String desc, int price) {
+        this.id = id;
+        this.name = name;
+        this.desc = desc;
+        this.price = price;
+    }
+
+    public Theme(String name, String desc, int price) {
+        this.name = name;
+        this.desc = desc;
+        this.price = price;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getDesc() {
+        return desc;
+    }
+
+    public int getPrice() {
+        return price;
+    }
+}
