+++ src/main/resources/init.sql
@@ -3,7 +3,11 @@ TRUNCATE TABLE RESERVATION;
 TRUNCATE TABLE SCHEDULE;
 TRUNCATE TABLE THEME;
 
-ALTER TABLE MEMBER ALTER COLUMN ID RESTART WITH 1;
-ALTER TABLE RESERVATION ALTER COLUMN ID RESTART WITH 1;
-ALTER TABLE SCHEDULE ALTER COLUMN ID RESTART WITH 1;
-ALTER TABLE THEME ALTER COLUMN ID RESTART WITH 1;
+ALTER TABLE MEMBER
+    ALTER COLUMN ID RESTART WITH 1;
+ALTER TABLE RESERVATION
+    ALTER COLUMN ID RESTART WITH 1;
+ALTER TABLE SCHEDULE
+    ALTER COLUMN ID RESTART WITH 1;
+ALTER TABLE THEME
+    ALTER COLUMN ID RESTART WITH 1;
