+++ src/main/resources/schema.sql
@@ -1,8 +1,8 @@
 CREATE TABLE reservation
 (
-    id          bigint      not null auto_increment,
-    schedule_id bigint      not null,
-    member_id   bigint      not null,
+    id          bigint not null auto_increment,
+    schedule_id bigint not null,
+    member_id   bigint not null,
     primary key (id)
 );
